+++ src/main/java/timelessodyssey/model/game/elements/Spike.java
@@ -1,7 +1,7 @@
 package timelessodyssey.model.game.elements;
 
 public class Spike extends Element{
-    public Spike(int x, int y) {
+    public Spike(double x, double y) {
         super(x, y);
     }
 }
